Methocode.com is finally alive!!!
After many searches I finally managed to route the domain and the Ip. Now we are up and running. However there is still a lot of things to do such as implementing our wireframe design into the website. We have to learn about pathing and serving static files in the node backend. However we managed to get something going.

Now let me introduce to you version 0.1.0a. We have a lot of things to do so let's get to it.